requirements.txt
fastapi
uvicorn[standard]
aiomysql
python-dotenv
pydantic